Change credits type from int to double

+++ staryu.cpp
@@ -27,7 +27,7 @@ struct Point {
 // Major Statistics
 struct stat {
     string interest;
-    int credits;
+    double credits;
 };
 
 
@@ -148,7 +148,7 @@ int main () {
     svg << "<polygon class=\"star-shape\" points=\"";
     for (int i=0; i < n; ++i) {
         double angle = i * (2 * PI / n) - PI / 2;
-        int clamped_credits = clamp(stats[i].credits, 0.0, maxVal); // clamp credit (0-maxValue)
+        double clamped_credits = clamp(stats[i].credits, 0.0, maxVal); // clamp credit (0-maxValue)
         Point p = getCoordinates(clamped_credits, angle, center, radius, maxVal);
         svg << p.x << "," << p.y << " ";
     }
